Table
you first need to set the
style for your table. You can use this example below as a reference for the code structure. The <td>
element is used to define a data cell within a table, <th> element is used to define a header cell within
a table and <tr> element is used to define a row within a table